% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsets.r
\docType{data}
\name{almond}
\alias{almond}
\title{River Almond invertebrate dataset}
\format{DataFrame}
\description{
The dataset is part of that derived from samples taken by
David Lawson as part of his undergraduate research project in Spring 2015.
Samples were collected using standard three-minute kick sampling techniques
from sites on the River Almond, West Lothian, Scotland above and below
significant weirs.
}
\author{
David Lawson, 2015
}
\keyword{datasets}

