% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMWPtab.r
\docType{data}
\name{BMWPtab}
\alias{BMWPtab}
\title{Table of biotic index scores}
\format{DataFrame}
\usage{
BMWPtab
}
\description{
Full list of scores used in calculations
}
\author{
Rob Briers, 2016
}
\keyword{datasets}

