% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsets.r
\docType{data}
\name{braidburn}
\alias{braidburn}
\title{Braid Burn invertebrate dataset}
\format{DataFrame}
\description{
The dataset is part of that derived from samples taken by
the Edinburgh Napier MSc Wildlife Biology and Conservation group in
Spring 2015.
Samples were collected using standard three-minute kick sampling
techniques from sites on the Braid Burn within the Hermitage of
Braid Local Nature Reserve, Edinburgh. The data are recorded as alphabetic
log abundance categories (A-E).
}
\author{
Rob Briers and MSc Wildlife Biology and Conservation students, 2015
}
\keyword{datasets}

