% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{calcAWIC}
\alias{calcAWIC}
\title{Calculate AWIC index}
\usage{
calcAWIC(df, type = "num")
}
\arguments{
\item{df}{A dataframe containing list of taxa in first column, followed by
columns of abundances with sample names in header row.}

\item{type}{Indicates format of data. Options are "num" for numeric data,
"log" for integer log abundance categories (1-5) or "alpha" for alphabetic
abundance categories (A-E). Default value is "num".}
}
\value{
A data frame consisting of a column of index values with samples
in rows.
}
\description{
Calculates AWIC index for invertebrate samples.
}
\examples{

# calculate the AWIC index for the Green Burn dataset
# data are numeric log abundance categories, so type is "log"

calcAWIC(greenburn, "log")
}

