% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcindex.r
\name{calcindex}
\alias{calcindex}
\title{Calculate biotic indices for invertebrate samples.}
\usage{
calcindex(df, index = "BMWP", tidy = FALSE)
}
\arguments{
\item{df}{A dataframe containing list of taxa followed by their abundances
in samples, with sample names as headers.  Default format is for taxon
names to be in the first column and sample abundances in subsequent
columns, but data can also be processed in the transposed layout by
setting the \code{tidy} argument to TRUE (see below).}

\item{index}{A choice of index to be calculated. Defaults to BMWP.
Options are: \code{"BMWP"}, \code{"Whalley"}, \code{"Riffle"}, \code{"Pool"},
\code{"RiffPool"}, \code{"LIFE"}, \code{"PSI"}, \code{"WHPT"},
\code{"WHPT_AB"} and \code{"AWIC"}.}

\item{tidy}{Whether the data are in a 'tidy' format (sensu Wickham). The
default is FALSE as data are commonly stored as columns as samples rather
than rows (with the taxon list in the first column). This is the default
format; if set to FALSE the data are transposed prior to processing.}
}
\value{
A data frame consisting of columns of indices with samples in rows
 (tidy format). The number of columns returned depends on the index
 selected.
}
\description{
Calculates a range of freshwater invertebrate biotic indices
 in use in the UK (based on family level identification. Currently
 calculates BMWP (N-taxa and ASPT), Whalley revised BMWP, Whalley habitat
  specific BWMP (riffle, pool and riffle/pool), LIFE, PSI, WHPT (presence
  -absence and abundance-weighted) and AWIC.  For details of these indices
   see the vignette.
}
\examples{
# load the built-in River Almond dataset

data(almond)

# calculate the PSI index for this dataset
# tidy is not specified as default is used (FALSE)

calcindex(almond, index="PSI")
}

