% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcscore.r
\name{calcscore}
\alias{calcscore}
\title{Calculates biotic index scores for individual samples.}
\usage{
calcscore(abundances, taxonlist, index)
}
\arguments{
\item{abundances}{An integer vector representing the abundances of taxa
in a sample. Absences should be represented as \code{NA} rather than
zero values but it will process data with zeros.}

\item{taxonlist}{A string vector containing taxon names (family level)
present in all the samples being analysed.}

\item{index}{A string representing the choice of index value. Defaults to
BMWP.}
}
\value{
A list containing the different components of each score. List
dimensions depend on the index calculated.
}
\description{
Calculates chosen biotic index on individual samples,
based on vectors of abundances and taxon names passed as arguments.
}

