% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.r
\name{checktaxa}
\alias{checktaxa}
\title{Check taxa against scoring list}
\usage{
checktaxa(df)
}
\arguments{
\item{df}{A dataframe containing abundances of invertebrate taxa in
different samples.}
}
\value{
A data frame containing the names of taxa that are not in the
list of scoring taxa, or \code{NA} if all taxa are scoring.
}
\description{
Check the list of taxa present in the sample dataset against
the list of scoring taxa within package to identify any non-scoring taxa
in the samples (or spelling mistakes).
}

