% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.r
\name{convertalpha}
\alias{convertalpha}
\title{Convert alphabetic log abundance classes.}
\usage{
convertalpha(df)
}
\arguments{
\item{df}{A dataframe containing abundances of invertebrate taxa in
different samples recorded as alphabetic categories (character format).}
}
\value{
A data frame with alphabetic categories converted to
appropriate numeric values for index calculation.
}
\description{
Takes data recorded as alphabetic log abundance classes
and converts them to numeric format, with abritary abundances within
each class: A=1, B=10, C=100, D=1000, E=1000.
}

