% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.r
\name{extractrows}
\alias{extractrows}
\title{Extract taxon scores based on presence.}
\usage{
extractrows(df, indextable)
}
\arguments{
\item{df}{A dataframe or atomic vector representing the sample. If in vector
form, should be a string vector of taxon names. If a dataframe, should
contain taxon names in strong format in first column, followed by columns
of numeric abundances in samples.}

\item{indextable}{A dataframe of biotic index values, with strings of taxon
names in the first column and scores in subsequent columns. Usually
loaded from object in the sysdata.rda file.}
}
\value{
A dataframe consisting of the rows of the index table matching the
taxa present in the sample.
}
\description{
Extracts a subset of biotic index score table data based on
taxon names in first row of dataframe passed in which represent taxa
present in any of the samples being analysed.
}

