% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.r
\name{findcomposites}
\alias{findcomposites}
\title{Locates BMWP composite families in sample data}
\usage{
findcomposites(composites, df)
}
\arguments{
\item{composites}{A dataframe containing the list of composite taxa with
each pair of composites in two columns.}

\item{df}{A dataframe containing list of taxa in first column, followed by
columns of abundances with sample names in header row.}
}
\value{
An indexing vector, giving the rows of the sample data that need
to be deleted to avoid double counting.
}
\description{
Takes a listing of taxa and abundances and locates those
taxa that are represented as composites in the BMWP index to avoid double
counting of scores.
}

