% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsets.r
\docType{data}
\name{greenburn}
\alias{greenburn}
\title{Green Burn invertebrate dataset}
\format{DataFrame}
\description{
The dataset is derived from samples taken by the Edinburgh
Napier MSc Aquatic Ecosystem Management group in Autumn 2012.
Samples were collected using standard three-minute kick sampling techniques
from a site on the Green Burn, a tributary of the River Almond in West
Lothian, Scotland. The data are recorded in numeric log abundance
categories (1-5).
}
\author{
Rob Briers and MSc Aquatic Ecosystem Management students, 2012
}
\keyword{datasets}

