% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.r
\name{logslice}
\alias{logslice}
\title{Classify taxa into log class score sets based on their sample abundance.}
\usage{
logslice(df)
}
\arguments{
\item{df}{A dataframe containing list of taxa in first column, followed by a column of abundances with sample names in header row.}
}
\value{
A data frame consisting of taxon names in the first column followed by all taxon index values and a log abundance class index column.
}
\description{
Takes a dataframe of taxon names and abundances and returns a dataframe of taxon index scores (through a call to extractrows), along with a 'class' index column indicating the log abundance class for each taxon in the sample.
}

