% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{calcAWIC}
\alias{calcAWIC}
\title{Calculates AWIC score indices for invertebrate samples.}
\usage{
calcAWIC(df, tidy = FALSE)
}
\arguments{
\item{df}{A dataframe containing list of taxa in first column, followed by
columns of abundances with sample names in header row.}

\item{tidy}{Whether the data are in a 'tidy' format (sensu Wickham). The
default is FALSE as data are commonly stored as columns as samples rather
than rows (with the taxon list in the first column). This is the default
 format; if set to TRUE the data are transposed prior to processing using
 the internal \code{\link{transposedata}} function.}
}
\value{
A data frame consisting of columns of indices with samples in
rows (tidy format).
}
\description{
Calculates AWIC score indices for invertebrate samples.
}

